move web implementation from cmd/api to the server package
rename the command api to serve
